moved add and edit products routes to admin and updated
